Split html/garden.rs into focused submodules
Replace the ~2,100-line garden.rs with a garden/ directory: routes,
render, vote compare, pin HUD, item page view-model, external resolvers,
breadcrumbs, access helpers, and unit tests. Public exports from mod.rs
are unchanged.
